requirements.txt
Flask
openpyxl
pywebview
azure-ai-documentintelligence
python-dotenv
